merge from branch glb-opt, pass float test

+++ src/structure/value.hh
@@ -13,7 +13,7 @@ class CTValue: public Info {
 public:
     DeclType type;
     int32_t int_value;
-    float float_value;
+    double float_value;
 public:
     CTValue(DeclType t = TypeVoid, int32_t iv = 0, float fv = 0);
     string ToString();
